Add AbstractPulsarClient
